requirements.txt
livereload==2.7.1
tornado==6.5.2
google-genai>=1.54.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
Pillow>=10.4.0
python-pptx>=0.6.23
